Consolidate users and hosts api into single file
