+++ pages/api/users.ts
@@ -1,58 +1,52 @@
 import axios from 'axios';
+import { getSession } from 'next-auth/react';
 import { AIRTABLE_API_KEY, AIRTABLE_URL } from '../../config';
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { cities } from '../../data/citiesData';
+import { filterByFormula } from '../../utils';
 
-const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  if (req.method === 'POST') {
-    console.log('Called with', req.body);
-    const {
-      name,
-      userType,
-      phoneNumber,
-      email,
-      city,
-      accomodationDetails,
-      groupSize,
-      languages,
-      dateStart,
-      dateEnd,
-      avatar,
-    } = req.body;
-    try {
-      const response = await axios({
-        method: 'POST',
-        url: `${AIRTABLE_URL}/Hosts`,
-        data: {
-          records: [
-            {
-              fields: {
-                name,
-                userType,
-                phoneNumber,
-                email,
-                city,
-                accomodationDetails,
-                groupSize,
-                languages,
-                dateStart,
-                dateEnd,
-                avatar,
-                ...cities[city as keyof typeof cities],
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
+  try {
+    const session = await getSession({ req });
+
+    switch (req.method) {
+      case 'GET':
+        const { data: user } = await axios({
+          method: 'GET',
+          url: `${AIRTABLE_URL}/Hosts?maxRecords=1&${filterByFormula(
+            'email',
+            session!.user!.email!
+          )}`,
+          headers: {
+            Authorization: `Bearer ${AIRTABLE_API_KEY}`,
+          },
+        });
+        return res.status(200).json({ user: user.records[0]?.fields });
+
+      case 'POST':
+        const { termsOfService, ...rest } = req.body;
+        const { data: created } = await axios({
+          method: 'POST',
+          url: `${AIRTABLE_URL}/Hosts`,
+          data: {
+            records: [
+              {
+                fields: rest,
               },
-            },
-          ],
-          typecast: true,
-        },
-        headers: { Authorization: `Bearer ${AIRTABLE_API_KEY}` },
-      });
-      res.status(200).json({ created: response.data });
-    } catch (err) {
-      res.status(500).json({ error: err });
-    }
-  } else {
-    res.status(404).json({ info: 'method not implemented' });
-  }
-};
+            ],
+            typecast: true,
+          },
+          headers: { Authorization: `Bearer ${AIRTABLE_API_KEY}` },
+        });
+        return res.status(200).json({ created });
 
-export default handler;
+      default:
+        res.status(404).json({ info: 'method not implemented' });
+    }
+  } catch (error: any) {
+    console.log(error.message);
+    res.status(500).json({ error: error.message });
+  }
+}
